Added config to block the breaking of vehicles if you hit walls and hard blocks.  Implements #707.

+++ src/main/java/minecrafttransportsimulator/vehicles/main/EntityVehicleC_Colliding.java
@@ -465,7 +465,7 @@ abstract class EntityVehicleC_Colliding extends EntityVehicleB_Rideable{
 				}
 			}
 			
-			if(hardnessHitThisTick > currentMass/(0.75 + velocity)/250F){
+			if(ConfigSystem.configObject.general.vehicleDestruction.value && hardnessHitThisTick > currentMass/(0.75 + velocity)/250F){
 				if(!world.isClient()){
 					destroyAt(box.globalCenter);
 				}
